총알 위치 문제
- 마우스 커서의 위치가 클라이언트 윈도우가 아닌 전체화면 기준으로 받아오는 문제 해결

+++ tank/Network.cpp
@@ -176,8 +176,8 @@ void do_recv()
 			
 			int pos_x = packet->pos_x;
 			int pos_y = packet->pos_y;
-			int aim_x = packet->aim_x;
-			int aim_y = packet->aim_y;
+			float aim_x = packet->aim_x;
+			float aim_y = packet->aim_y;
 
 			mPlayer[p_id].x = pos_x;
 			mPlayer[p_id].y = pos_y;
